refactor(console): role users transfer component
