+++ packages/console/src/components/RoleUsersTransfer/components/SourceUsersBox/index.module.scss
@@ -1,0 +1,7 @@
+.search {
+  width: 100%;
+}
+
+.icon {
+  color: var(--color-text-secondary);
+}
