+++ packages/console/src/components/RoleUsersTransfer/components/TargetUsersBox/index.module.scss
@@ -1,0 +1,3 @@
+.added {
+  font: var(--font-label-large);
+}
