+++ packages/console/src/components/RoleUsersTransfer/index.module.scss
@@ -1,81 +1,5 @@
 @use '@/scss/underscore' as _;
 
-.container {
-  border: 1px solid var(--color-border);
-  border-radius: 6px;
-  display: flex;
-  align-items: stretch;
-  overflow: hidden;
+.roleUsersTransfer {
   height: 360px;
-
-  .verticalBar {
-    @include _.vertical-bar;
-  }
-
-  .box {
-    flex: 1 1 0;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-
-    .top {
-      height: 52px;
-      border-bottom: 1px solid var(--color-border);
-      display: flex;
-      align-items: center;
-      padding: 0 _.unit(4);
-
-      .search {
-        width: 100%;
-      }
-
-      &.added {
-        font: var(--font-label-large);
-      }
-    }
-
-    .content {
-      flex: 1 1 0;
-      overflow-y: auto;
-
-      .item {
-        display: flex;
-        align-items: center;
-        padding: _.unit(2.5) _.unit(4);
-        cursor: pointer;
-        user-select: none;
-
-        .avatar {
-          width: 20px;
-          height: 20px;
-          border-radius: 4px;
-          object-fit: cover;
-          flex-shrink: 0;
-          margin-right: _.unit(2);
-        }
-
-        .name {
-          flex: 1 1 0;
-          font: var(--font-body-medium);
-          @include _.text-ellipsis;
-        }
-
-        &:hover {
-          background: var(--color-hover);
-        }
-      }
-    }
-
-    .pagination {
-      height: 40px;
-      padding-right: _.unit(4);
-      border-top: 1px solid var(--color-border);
-    }
-
-    .icon {
-      color: var(--color-text-secondary);
-    }
-  }
 }
-
